feat: Project RPG
스킬 밸런싱

+++ 02_CSharpAlgorithm/PersonalProject/ProjectRPG/Skills/Passives/Skill_DeepBreath.cs
@@ -35,7 +35,7 @@
                 return false;
             if (!used)
             {
-                param1[1, 1] += value * (level + rank * 10);
+                param1[1, 1] += (int)value * (level + rank * 10);
                 used = true;
             }
             return true;
